Restructure again, touching up scripts and removing non-deterministic parts from flake
This commit moves a lot of the scripts that nix would run into
the root directory and out of nix's purview.

These scripts do things like create ssh key-pairs, so Nix would complain
that builds results could change without the derivation having changed.

These were moved out so that the environment created by Nix is the same,
regardless of where or when it is created.

This commit also changes the scripts in this repo to make logging more
uniform and does small touchups to them.